requirements.txt
mcp>=0.9.0
fastmcp>=0.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
tenacity>=8.2.0